+++ wdc/controller/calculator.py
@@ -1,7 +1,7 @@
 from wdc.time import WdcTime
 
 
-def calc_workday_end(start_time: str, break_duration: int, workday_duration: str):
+def calculate(start_time: str, break_duration: int, workday_duration: str):
     start = WdcTime(start_time)
     duration = WdcTime(workday_duration)